requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
autocuda==0.16
billiard==4.2.0
blinker==1.6.3
blis==0.7.11
boostaug==2.3.5
catalogue==2.0.10
celery==5.3.6
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.15.1
colorama==0.4.6
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.12.4
findfile==2.0.1
Flask==3.0.0
fonttools==4.43.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.20.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
kombu==5.3.4
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
metric-visualizer==0.9.9
mpmath==1.3.0
murmurhash==1.0.10
natsort==8.4.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pathy==0.10.3
Pillow==10.0.1
pipreqs==0.4.13
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==3.20.3
pyabsa==2.3.4
pydantic==2.4.2
pydantic_core==2.10.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic==0.4.27
pytorch-warmup==0.1.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.0
sentencepiece==0.1.99
seqeval==1.2.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
thinc==8.2.2
threadpoolctl==3.2.0
tikzplotlib==0.10.1
tokenizers==0.13.3
torch==2.1.2
tqdm==4.66.1
transformers==4.29.2
triton==2.1.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.7
vine==5.1.0
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
Werkzeug==3.0.1
wordcloud==1.9.3
XlsxWriter==3.1.9
yarg==0.1.9
zipp==3.17.0
